Crate rusoto_mediastore

An AWS Elemental MediaStore container is a namespace that holds folders and objects. You use a container endpoint to create, read, and delete objects.

If you're using the service, you're probably looking for MediaStoreClient and MediaStore.

Structs

Container
This section describes operations that you can perform on an AWS Elemental MediaStore container.

CorsRule
A rule for a CORS policy. You can add up to 100 rules to a CORS policy. If more than one rule applies, the service uses the first applicable rule listed.

CreateContainerInput

CreateContainerOutput

DeleteContainerInput

DeleteContainerOutput

DeleteContainerPolicyInput

DeleteContainerPolicyOutput

DeleteCorsPolicyInput

DeleteCorsPolicyOutput

DeleteLifecyclePolicyInput

DeleteLifecyclePolicyOutput

DescribeContainerInput

DescribeContainerOutput

GetContainerPolicyInput

GetContainerPolicyOutput

GetCorsPolicyInput

GetCorsPolicyOutput

GetLifecyclePolicyInput

GetLifecyclePolicyOutput

ListContainersInput

ListContainersOutput

MediaStoreClient
A client for the MediaStore API.

PutContainerPolicyInput

PutContainerPolicyOutput

PutCorsPolicyInput

PutCorsPolicyOutput

PutLifecyclePolicyInput

PutLifecyclePolicyOutput

StartAccessLoggingInput

StartAccessLoggingOutput

StopAccessLoggingInput

StopAccessLoggingOutput

Enums

CreateContainerError
Errors returned by CreateContainer

DeleteContainerError
Errors returned by DeleteContainer

DeleteContainerPolicyError
Errors returned by DeleteContainerPolicy

DeleteCorsPolicyError
Errors returned by DeleteCorsPolicy

DeleteLifecyclePolicyError
Errors returned by DeleteLifecyclePolicy

DescribeContainerError
Errors returned by DescribeContainer

GetContainerPolicyError
Errors returned by GetContainerPolicy

GetCorsPolicyError
Errors returned by GetCorsPolicy

GetLifecyclePolicyError
Errors returned by GetLifecyclePolicy

ListContainersError
Errors returned by ListContainers

PutContainerPolicyError
Errors returned by PutContainerPolicy

PutCorsPolicyError
Errors returned by PutCorsPolicy

PutLifecyclePolicyError
Errors returned by PutLifecyclePolicy

StartAccessLoggingError
Errors returned by StartAccessLogging

StopAccessLoggingError
Errors returned by StopAccessLogging

Traits

MediaStore
Trait representing the capabilities of the MediaStore API. MediaStore clients implement this trait.