Crate rusoto_mediastore

An AWS Elemental MediaStore container is a namespace that holds folders and objects. You use a container endpoint to create, read, and delete objects.

If you're using the service, you're probably looking for MediaStoreClient and MediaStore.

Structs

Container
This section describes operations that you can perform on an AWS Elemental MediaStore container.

CorsRule
A rule for a CORS policy. You can add up to 100 rules to a CORS policy. If more than one rule applies, the service uses the first applicable rule listed.

CreateContainerInput

CreateContainerOutput

DeleteContainerInput

DeleteContainerOutput

DeleteContainerPolicyInput

DeleteContainerPolicyOutput

DeleteCorsPolicyInput

DeleteCorsPolicyOutput

DescribeContainerInput

DescribeContainerOutput

GetContainerPolicyInput

GetContainerPolicyOutput

GetCorsPolicyInput

GetCorsPolicyOutput

ListContainersInput

ListContainersOutput

MediaStoreClient
A client for the MediaStore API.

PutContainerPolicyInput

PutContainerPolicyOutput

PutCorsPolicyInput

PutCorsPolicyOutput

Enums

CreateContainerError
Errors returned by CreateContainer

DeleteContainerError
Errors returned by DeleteContainer

DeleteContainerPolicyError
Errors returned by DeleteContainerPolicy

DeleteCorsPolicyError
Errors returned by DeleteCorsPolicy

DescribeContainerError
Errors returned by DescribeContainer

GetContainerPolicyError
Errors returned by GetContainerPolicy

GetCorsPolicyError
Errors returned by GetCorsPolicy

ListContainersError
Errors returned by ListContainers

PutContainerPolicyError
Errors returned by PutContainerPolicy

PutCorsPolicyError
Errors returned by PutCorsPolicy

Traits

MediaStore
Trait representing the capabilities of the MediaStore API. MediaStore clients implement this trait.