Struct rusoto_mediastore::MediaStoreClient

pub struct MediaStoreClient { /* fields omitted */ }

A client for the MediaStore API.

Methods

impl MediaStoreClient

pub fn new(region: Region) -> MediaStoreClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MediaStoreClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,