Trait rusoto_mediastore::MediaStore

pub trait MediaStore {
    fn create_container(
        &self, 
        input: CreateContainerInput
    ) -> RusotoFuture<CreateContainerOutput, CreateContainerError>;

    fn delete_container(
        &self, 
        input: DeleteContainerInput
    ) -> RusotoFuture<DeleteContainerOutput, DeleteContainerError>;

    fn delete_container_policy(
        &self, 
        input: DeleteContainerPolicyInput
    ) -> RusotoFuture<DeleteContainerPolicyOutput, DeleteContainerPolicyError>;

    fn delete_cors_policy(
        &self, 
        input: DeleteCorsPolicyInput
    ) -> RusotoFuture<DeleteCorsPolicyOutput, DeleteCorsPolicyError>;

    fn describe_container(
        &self, 
        input: DescribeContainerInput
    ) -> RusotoFuture<DescribeContainerOutput, DescribeContainerError>;

    fn get_container_policy(
        &self, 
        input: GetContainerPolicyInput
    ) -> RusotoFuture<GetContainerPolicyOutput, GetContainerPolicyError>;

    fn get_cors_policy(
        &self, 
        input: GetCorsPolicyInput
    ) -> RusotoFuture<GetCorsPolicyOutput, GetCorsPolicyError>;

    fn list_containers(
        &self, 
        input: ListContainersInput
    ) -> RusotoFuture<ListContainersOutput, ListContainersError>;

    fn put_container_policy(
        &self, 
        input: PutContainerPolicyInput
    ) -> RusotoFuture<PutContainerPolicyOutput, PutContainerPolicyError>;

    fn put_cors_policy(
        &self, 
        input: PutCorsPolicyInput
    ) -> RusotoFuture<PutCorsPolicyOutput, PutCorsPolicyError>;
}

Trait representing the capabilities of the MediaStore API. MediaStore clients implement this trait.

Required Methods

fn create_container(
&self,
input: CreateContainerInput
) -> RusotoFuture<CreateContainerOutput, CreateContainerError>

Creates a storage container to hold objects. A container is similar to a bucket in the Amazon S3 service.

fn delete_container(
&self,
input: DeleteContainerInput
) -> RusotoFuture<DeleteContainerOutput, DeleteContainerError>

Deletes the specified container. Before you make a DeleteContainer request, delete any objects in the container or in any folders in the container. You can delete only empty containers.

fn delete_container_policy(
&self,
input: DeleteContainerPolicyInput
) -> RusotoFuture<DeleteContainerPolicyOutput, DeleteContainerPolicyError>

Deletes the access policy that is associated with the specified container.

fn delete_cors_policy(
&self,
input: DeleteCorsPolicyInput
) -> RusotoFuture<DeleteCorsPolicyOutput, DeleteCorsPolicyError>

Deletes the cross-origin resource sharing (CORS) configuration information that is set for the container.

To use this operation, you must have permission to perform the MediaStore:DeleteCorsPolicy action. The container owner has this permission by default and can grant this permission to others.

fn describe_container(
&self,
input: DescribeContainerInput
) -> RusotoFuture<DescribeContainerOutput, DescribeContainerError>

Retrieves the properties of the requested container. This request is commonly used to retrieve the endpoint of a container. An endpoint is a value assigned by the service when a new container is created. A container's endpoint does not change after it has been assigned. The DescribeContainer request returns a single Container object based on ContainerName. To return all Container objects that are associated with a specified AWS account, use ListContainers.

fn get_container_policy(
&self,
input: GetContainerPolicyInput
) -> RusotoFuture<GetContainerPolicyOutput, GetContainerPolicyError>

Retrieves the access policy for the specified container. For information about the data that is included in an access policy, see the AWS Identity and Access Management User Guide.

fn get_cors_policy(
&self,
input: GetCorsPolicyInput
) -> RusotoFuture<GetCorsPolicyOutput, GetCorsPolicyError>

Returns the cross-origin resource sharing (CORS) configuration information that is set for the container.

To use this operation, you must have permission to perform the MediaStore:GetCorsPolicy action. By default, the container owner has this permission and can grant it to others.

fn list_containers(
&self,
input: ListContainersInput
) -> RusotoFuture<ListContainersOutput, ListContainersError>

Lists the properties of all containers in AWS Elemental MediaStore.

You can query to receive all the containers in one response. Or you can include the MaxResults parameter to receive a limited number of containers in each response. In this case, the response includes a token. To get the next set of containers, send the command again, this time with the NextToken parameter (with the returned token as its value). The next set of responses appears, with a token if there are still more containers to receive.

See also DescribeContainer, which gets the properties of one container.

fn put_container_policy(
&self,
input: PutContainerPolicyInput
) -> RusotoFuture<PutContainerPolicyOutput, PutContainerPolicyError>

Creates an access policy for the specified container to restrict the users and clients that can access it. For information about the data that is included in an access policy, see the AWS Identity and Access Management User Guide.

For this release of the REST API, you can create only one policy for a container. If you enter PutContainerPolicy twice, the second command modifies the existing policy.

fn put_cors_policy(
&self,
input: PutCorsPolicyInput
) -> RusotoFuture<PutCorsPolicyOutput, PutCorsPolicyError>

Sets the cross-origin resource sharing (CORS) configuration on a container so that the container can service cross-origin requests. For example, you might want to enable a request whose origin is http://www.example.com to access your AWS Elemental MediaStore container at my.example.container.com by using the browser's XMLHttpRequest capability.

To enable CORS on a container, you attach a CORS policy to the container. In the CORS policy, you configure rules that identify origins and the HTTP methods that can be executed on your container. The policy can contain up to 398,000 characters. You can add up to 100 rules to a CORS policy. If more than one rule applies, the service uses the first applicable rule listed.